API rework
Management of new features was becoming difficult. As more features were
added, more APIs needed to be called and correlated. Output types as a
result were starting to become complex as we wanted to bundle relevant
items together to save calls. Having an API call for every use case
isn't reasonable, and will become hard to manage.

Instead we now bundle all info in either a TvEpisode or a TvShow. These
types are no longer a reflection of what is in the DB, or what is in the
index. A show now contains it's watch status and pause state. The
TvEpisode now contains it's watch date.

Some of these changes result in more complex SQL queries than we used to
have, pushing some app logic into the DB.

Since items now contain their database index, we cannot create them
without going through the DB. That can be viewed as a nice constraint,
as we cannot present something to a user that we have not persisted yet.

With those changes, we can clean up the API to mirror a more familiar
REST API. Shows are under /shows, and individual shows can be grabbed
through /shows/id/, show episodes can be grabbed by /shows/id/episodes/.
Same for episodes. Modification of watch status, pause state, etc, can
be done by doing a PUT on the show/episode ID with the field in the
desired state. No attempt has been made to validate that only the inputs
we can handle have changed

The show page in the client has been reasonably re-worked as well. We
now render the page using dom nodes, instead of an html string that we
assign. This allows us to wrap the state of the page in a class that
caches the show/episode data, and assign onclick handlers to bind to the
class

Bonus changes:
* Errors that make it to the UI are now logged with their source chain
  correctly
* Unwraps are now forbidden by the lint script