Add movies to database
Movies have a lot in common with tv shows in that...
* I want to track which ones I want to see
* I want to know when a movie that I wanted to watch is out on dvd
* I want to know when a movie that I wanted to watch is in theaters
* I want to remember if I liked a movie or not

Add movie indexing to the backend. Movie metadata is retrieved through
the open movie database. This was chosen over TVDB and IMDB because...
* TVDB had a long terms of service that I didn't want to adhere to
* IMDB API seemed... complicated/expensive (with little research).
* OMDB has open in the name ;)

OMDB API key is now provided on the command line, and is required for
the app to work. This constraint could be relaxed in the future, but
AFAIK, I am the only user, and I have an OMDB API key.

Movies are fetched from OMDB, and inserted into the database in a
similar style to tv shows. Images are put into the image table,
ratings/watch status are in their own tables, the movie table is
strictly for what was fetched from the indexer.

Episode watch status table was renamed from watch_status as this is now
ambiguous

As part of this change, the image cache was split out of the mutex
protection in the App class. This is because the database is really the
only thing that needs to be locked here. This became apparent when
adding a new member that also did not need to be locked. A lot of
changes in App are moving around the locking of inner -> db. This
could/should have been a different commit, but it's not.

Add REST APIs to add a movie, get movies, get a specific movie, update
watch status/rating of a movie, and delete a movie.