Fix mismatched image ids and cached images
Previously we were using image IDs as the filename for images in ou
image cache. This was a bad idea, as if we switched databases or image
ids were reused due to deletion from the database, we would be caching
the wrong image at that ID

Swap ids for urlencoded urls (so that we can store them on the
filesystem), this completely dodges the problem.