Add feature to skip episodes instead of watch them
In some cases I may want to start a tv show half way through. E.g.
for reality tv you can get away with only watching the latest seasons.

In this scenario I need a way to tell the system "yes I haven't watched
this, no I never want to". This allows for a "completion" of a show
without having to lie about watching it

Add support for skipping shows

* DB has a new skipped episode table. Application is responsible for
  ensuring that we only have one of watched/skipped
* Watch status is now an enum instead of an optional date
* UI now has the ability to show/set a more complicated watch state
  * Replace hyperlink for watched/unwatched with a set of buttons for
    unwatched/skipped/watched for each episode
* UI now utilizes the skip state when determining which shows have new
  episodes to watch
* UI shows skipped state as orange in card showing watch state for a
  show at a glance