Fix inconsistent next up state
If at 11pm in vancouver, you mark all episodes of a show as watched, and
an episode will air tomorrow, the next up segment will still list the
show.

This is a combination of 2 bugs

1. The show page assumes the episode will air at 00:00:00 UTC time, when
   the server assumes that the episode will air at 00:00:00 localtime
2. We check for equality on episode watch status

1. results in us marking more episodes as watched than have aired, 2.
results in the watch list thinking the show is incomplete

Fix both of these issues. Use local date in the show page, and do a >=
check on how many episodes we've watched.

There are still multiple issues with the current state that I do not
feel like fixing now

* Timezone that the show will air in is not accounted for
* If we mark unaired episodes as watched, and we have multiple
  playthroughs, we'll end the additional playthroughs too early